+++ InfinitePay/Responses/TransactionDetailsResponse.cs
@@ -11,9 +11,9 @@ namespace InfinitePay.Responses{
         public string event_reference_id { get; set; }
         public int installments { get; set; }
         public List<InstallmentsDetailResponse> installments_details { get; set; }
-        public int amount { get; set; }
-        public double net_amount { get; set; }
-        public double tax_amount { get; set; }
+        public decimal amount { get; set; }
+        public decimal net_amount { get; set; }
+        public decimal tax_amount { get; set; }
         public DateTime created_at { get; set; }
         public BuyerDetailsResponse buyer_details { get; set; }
