+++ InfinitePay/Responses/TransactionResponse.cs
@@ -10,8 +10,8 @@ namespace InfinitePay.Responses{
         public string nsu { get; set; }
         public string event_reference_id { get; set; }
         public int installments { get; set; }
-        public int amount { get; set; }
-        public int net_amount { get; set; }
+        public decimal amount { get; set; }
+        public decimal net_amount { get; set; }
         public DateTime created_at { get; set; }
         public BuyerDetailsResponse BuyerDetailsResponse { get; set; }
     }
